Move tests for loopdiagram.py into tests directory
