+++ eppy/useful_scripts/loopdiagram.py
@@ -84,41 +84,6 @@ def dropnodes(edges):
     return newedges
     
     
-def test_dropnodes():
-    """py.test for dropnodes"""
-    # test 1
-    node = "node"
-    (a,b,c,d,e,f,g,h,i) = (('a', node),'b',('c', node),'d',
-        ('e', node),'f',('g', node),'h',('i', node))
-    edges = [(a, b),
-    (b, c),
-    (c, d),
-    (d, e),
-    (e, f),
-    (f, g),
-    (g, h),
-    (h, i),]
-    theresult = [('a', 'b'), ('b', 'd'), ('d', 'f'), ('f', 'h'), ('h', 'i')]
-    result = dropnodes(edges)
-    assert result == theresult
-    # test 2
-    (a,b,c,d,e,f,g,h,i,j) = (('a', node),'b',('c', node),
-        ('d', node),'e','f',('g', node),('h', node),'i',('j', node))
-    edges = [(a, b),
-    (b, c),
-    (c, e),
-    (e, g),
-    (g, i),
-    (i, j),
-    (b, d),
-    (d, f),
-    (f, h),
-    (h, i),]
-    theresult = [('a', 'b'), ('b', 'e'), ('e', 'i'), ('i', 'j'), 
-            ('b', 'f'), ('f', 'i')]
-    result = dropnodes(edges)
-    assert result == theresult
-    
 def makeanode(name):
     return pydot.Node(name, shape="plaintext", label=name)
     
@@ -152,15 +117,6 @@ def edges2nodes(edges):
     justnodes = sorted(justnodes, key=lambda x: str(x[0]))
     return justnodes
     
-def test_edges2nodes():
-    """py.test for edges2nodes"""
-    thedata = (([("a", "b"), ("b", "c"), ("c", "d")],
-    ["a", "b", "c", "d"]), # edges, nodes
-    )
-    for edges, nodes in thedata:
-        result = edges2nodes(edges)   
-        assert result == nodes
-        
 
 def makediagram(edges):
     """make the diagram with the edges"""
@@ -544,27 +500,6 @@ def clean_edges(arg):
     except TypeError: # catch when for loop fails
         return replace_colon(arg) # not a sequence so just return repr
 
-# start pytests +++++++++++++++++++++++
-
-def test_replace_colon():
-    """py.test for replace_colon"""
-    data = (("zone:aap", '@', "zone@aap"),# s, r, replaced
-    )    
-    for s, r, replaced in data:
-        result = replace_colon(s, r)
-        assert result == replaced
-        
-def test_cleanedges():
-    """py.test for cleanedges"""
-    data = (([('a:a', 'a'), (('a', 'a'), 'a:a'), ('a:a', ('a', 'a'))],
-    (('a__a', 'a'), (('a', 'a'), 'a__a'), ('a__a', ('a', 'a')))), 
-    # edg, clean_edg
-    )
-    for edg, clean_edg in data:
-        result = clean_edges(edg)
-        assert result == clean_edg
-        
-# end pytests +++++++++++++++++++++++
     
 def main():
     from argparse import RawTextHelpFormatter
